+++ packages/vue/src/slider/slider-marker-group.tsx
@@ -1,11 +1,11 @@
 import { defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
-import { getValidChildren, type ComponentWithProps } from '../utils'
+import { getValidChildren } from '../utils'
 import { useSliderContext } from './slider-context'
 
 export type SliderMarkerGroupProps = HTMLArkProps<'div'>
 
-export const SliderMarkerGroup: ComponentWithProps<SliderMarkerGroupProps> = defineComponent({
+export const SliderMarkerGroup = defineComponent({
   name: 'SliderMarkerGroup',
   setup(_, { slots, attrs }) {
     const api = useSliderContext()