+++ packages/vue/src/slider/use-slider.ts
@@ -1,38 +1,40 @@
-import { connect, machine } from '@zag-js/slider'
-import { normalizeProps, useMachine } from '@zag-js/vue'
-import { computed, reactive, type ExtractPropTypes, type UnwrapRef } from 'vue'
+import * as slider from '@zag-js/slider'
+import { normalizeProps, useMachine, type PropTypes } from '@zag-js/vue'
+import { computed, type ComputedRef } from 'vue'
 import { useEnvironmentContext } from '../environment'
-import { useId } from '../utils'
-import type { SliderContext } from './slider'
+import type { Optional } from '../types'
+import { generateEventMap, useId } from '../utils'
+import { emits } from './slider.props'
 
-export const useSlider = <T extends ExtractPropTypes<SliderContext>>(
-  emit: CallableFunction,
-  context: T,
-) => {
-  const reactiveContext = reactive(context)
+export type UseSliderProps = Optional<slider.Context, 'id'> & {
+  modelValue?: slider.Context['value']
+}
+export type UseSliderReturn = ComputedRef<slider.Api<PropTypes>>
 
+export const useSlider = (props: UseSliderProps, emit: CallableFunction): UseSliderReturn => {
   const getRootNode = useEnvironmentContext()
+  const context = computed(() => {
+    const { modelValue, ...rest } = props
+    return {
+      ...rest,
+      value: modelValue,
+    }
+  })
+  const eventMap = generateEventMap(emits, emit)
 
   const [state, send] = useMachine(
-    machine({
-      ...reactiveContext,
-      id: reactiveContext.id || useId().value,
+    slider.machine({
+      ...context.value,
+      id: context.value.id ?? useId().value,
       getRootNode,
-      value: reactiveContext.modelValue ?? reactiveContext.value,
-      onChange(details) {
-        emit('change', details)
+      ...eventMap,
+      onChange: (details) => {
+        emit('change', details.value)
         emit('update:modelValue', details.value)
       },
-      onChangeEnd(details) {
-        emit('change-end', details)
-      },
-      onChangeStart(details) {
-        emit('change-start', details)
-      },
     }),
+    { context },
   )
 
-  return computed(() => connect(state.value, send, normalizeProps))
+  return computed(() => slider.connect(state.value, send, normalizeProps))
 }
-
-export type UseSliderReturn = UnwrapRef<ReturnType<typeof useSlider>>