+++ packages/vue/CHANGELOG.md
@@ -22,6 +22,7 @@ description: All notable changes to this project will be documented in this file
 - Revised `NumberInput` component
 - Revised `Pagination` component
 - Revised `Select` component
+- Revised `Slider` component
 - Revised `Splitter` component
 - Revised `Tabs` component
 - Revised `Tooltip` component